% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Body Mass Index}
\usage{
bmi(height, weight)
}
\arguments{
\item{height}{height in inches}

\item{weight}{weight in lbs}
}
\value{
a numeric vector of bmi
}
\description{
Given weight and height, return the body mass index
}
\examples{
# The bmi of a six foot one inch tall, three hundred pound individual is
bmi(height = 6 * 12 + 1, 300)

}
